move callbacks to hook
