+++ packages/slate-plugins/src/elements/mention/useMention.tsx
@@ -1,18 +1,7 @@
-import React, { useState } from 'react';
-import { Editor, Range } from 'slate';
+import React, { useCallback, useState } from 'react';
+import { Editor, Range, Transforms } from 'slate';
 import { MentionSelect } from './components/MentionSelect';
-import { onChangeMention } from './onChangeMention';
-import { onKeyDownMention } from './onKeyDownMention';
-import { OnKeyDownMentionOptions } from './types';
-
-const createOnChangeMention = (
-  setTarget: (target: Range | null) => void,
-  setSearch: (term: string) => void,
-  setIndex: (i: number) => void
-) => {
-  return (editor: Editor, beforeRegex?: RegExp) =>
-    onChangeMention({ editor, setTarget, setSearch, setIndex, beforeRegex });
-};
+import { insertMention } from './transforms';
 
 const getMentionSelect = (
   target: Range | null,
@@ -25,10 +14,6 @@ const getMentionSelect = (
     }
     return null;
   };
-};
-
-const createOnKeyDownMentions = (options: OnKeyDownMentionOptions) => {
-  return () => onKeyDownMention(options);
 };
 
 export const useMention = ({
@@ -45,19 +30,77 @@ export const useMention = ({
     .filter((c) => c.toLowerCase().includes(search.toLowerCase()))
     .slice(0, maxSuggestions);
 
-  const defaultOnChangeMention = createOnChangeMention(
-    setTarget,
-    setSearch,
-    setIndex
-  );
   const MentionSelectComponent = getMentionSelect(target, index, chars);
-  const defaultOnKeyDownMention = createOnKeyDownMentions({
-    chars,
-    target,
-    setTarget,
-    index,
-    setIndex,
-  });
+
+  const onKeyDownMention = useCallback(
+    (e: any, editor: Editor) => {
+      if (target) {
+        switch (e.key) {
+          case 'ArrowDown': {
+            e.preventDefault();
+            const prevIndex = index >= chars.length - 1 ? 0 : index + 1;
+            setIndex(prevIndex);
+            break;
+          }
+          case 'ArrowUp': {
+            e.preventDefault();
+            const nextIndex = index <= 0 ? chars.length - 1 : index - 1;
+            setIndex(nextIndex);
+            break;
+          }
+          case 'Tab':
+          case 'Enter':
+            e.preventDefault();
+            Transforms.select(editor, target);
+            insertMention(editor, chars[index]);
+            setTarget(null);
+            break;
+          case 'Escape':
+            e.preventDefault();
+            setTarget(null);
+            break;
+          default:
+            break;
+        }
+      }
+    },
+    [chars, index, setIndex, target, setTarget]
+  );
+
+  const onChangeMention = useCallback(
+    ({
+      editor,
+      beforeRegex = /^@(\w+)$/,
+    }: {
+      editor: Editor;
+      beforeRegex?: RegExp;
+    }) => {
+      const { selection } = editor;
+
+      if (selection && Range.isCollapsed(selection)) {
+        const [start] = Range.edges(selection);
+        const wordBefore = Editor.before(editor, start, { unit: 'word' });
+        const before = wordBefore && Editor.before(editor, wordBefore);
+        const beforeRange = before && Editor.range(editor, before, start);
+        const beforeText = beforeRange && Editor.string(editor, beforeRange);
+        const beforeMatch = beforeText && beforeText.match(beforeRegex);
+        const after = Editor.after(editor, start);
+        const afterRange = Editor.range(editor, start, after);
+        const afterText = Editor.string(editor, afterRange);
+        const afterMatch = afterText.match(/^(\s|$)/);
+        if (beforeMatch && afterMatch) {
+          setTarget(beforeRange ?? null);
+          setSearch(beforeMatch[1]);
+          setIndex(0);
+          return;
+        }
+      }
+
+      setTarget(null);
+    },
+    [setTarget, setSearch, setIndex]
+  );
+
   return {
     target,
     setTarget,
@@ -67,7 +110,7 @@ export const useMention = ({
     setSearch,
     chars,
     MentionSelectComponent,
-    defaultOnChangeMention,
-    defaultOnKeyDownMention,
+    onChangeMention,
+    onKeyDownMention,
   };
 };
