+++ stories/plugins/mention.stories.tsx
@@ -6,8 +6,6 @@ import {
   EditablePlugins,
   MentionPlugin,
   MentionSelect,
-  onChangeMention,
-  onKeyDownMention,
   useMention,
   withMention,
 } from '../../packages/slate-plugins/src';
@@ -19,8 +17,6 @@ export default {
   component: MentionPlugin,
   subcomponents: {
     useMention,
-    onChangeMention,
-    onKeyDownMention,
     MentionSelect,
   },
 };
@@ -36,7 +32,7 @@ export const Example = () => {
       []
     );
 
-    const { MentionSelectComponent,  defaultOnChangeMention, defaultOnKeyDownMention } = useMention(
+    const { MentionSelectComponent,  onChangeMention, onKeyDownMention } = useMention(
       {
         characters: CHARACTERS,
         maxSuggestions: 10,
@@ -50,14 +46,14 @@ export const Example = () => {
         onChange={(newValue) => {
           setValue(newValue);
 
-          defaultOnChangeMention(editor);
+          onChangeMention({editor});
         }}
       >
         <EditablePlugins
           plugins={plugins}
           placeholder="Enter some text..."
           onKeyDown={[
-            defaultOnKeyDownMention(),
+            onKeyDownMention,
           ]}
         />
           <MentionSelectComponent />
